requirements/local.txt
-r base.txt

django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework-stubs==3.16.7
types-pyyaml==6.0.12.20250915
types-requests==2.32.4.20260107
typing-extensions==4.15.0